refactor: Leagcy 코드 삭제
