+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -5,26 +5,43 @@ import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
+    @RequestMapping(value = "/login", method = RequestMethod.GET)
+    public ModelAndView show(HttpServletRequest req, HttpServletResponse res) {
+        String viewName = UserSession.getUserFrom(req.getSession())
+            .map(user -> {
+                log.info("logged in {}", user.getAccount());
+                return "redirect:/index.jsp";
+            })
+            .orElse("/login.jsp");
+        return new ModelAndView(new JspView(viewName));
+    }
+
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
+    public ModelAndView save(HttpServletRequest req, HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
-            return "redirect:/index.jsp";
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
         }
 
-        return InMemoryUserRepository.findByAccount(req.getParameter("account"))
-                .map(user -> {
-                    log.info("User : {}", user);
-                    return login(req, user);
-                })
-                .orElse("redirect:/401.jsp");
+        String viewName = InMemoryUserRepository.findByAccount(req.getParameter("account"))
+            .map(user -> {
+                log.info("User : {}", user);
+                return login(req, user);
+            })
+            .orElse("redirect:/401.jsp");
+        return new ModelAndView(new JspView(viewName));
     }
 
     private String login(HttpServletRequest request, User user) {
